% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{brier_score}
\alias{brier_score}
\title{Calculate the brier score}
\usage{
brier_score(time, status, predictYear, survival)
}
\arguments{
\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{predictYear}{Predicted t year of disease's risk}

\item{survival}{Predicted survival at time t}
}
\description{
Calculate the brier score
}
\keyword{internal}
