% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkNonlinearX}
\alias{checkNonlinearX}
\title{Check the non-linear association of predictor on the outcome}
\usage{
checkNonlinearX(dat, time, status, x, Yeoi)
}
\arguments{
\item{dat}{Data set used or the derivation cohort}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{x}{The continuous predictor of interest}

\item{Yeoi}{Event value of the outcome of interest}
}
\description{
Check the non-linear association of predictor on the outcome
}
