% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxRiskModels.R
\name{CoxModelBootstrapping}
\alias{CoxModelBootstrapping}
\title{Develop risk prediction models with internal validation using bootstrap method}
\usage{
CoxModelBootstrapping(
  dat,
  time,
  status,
  Xs,
  predictYear,
  KMS0 = FALSE,
  riskThreshold,
  calGroup,
  nboot = 100
)
}
\arguments{
\item{dat}{Data set or the derivation cohort}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{Xs}{Candidate continuous predictors used for developing risk prediction models.}

\item{predictYear}{The horizon year or t-year predicted risk}

\item{KMS0}{Whether calculate the baseline survival rate using the Kaplan-Meier estimator}

\item{riskThreshold}{User-defined thresholds used for calculating calibration measures}

\item{calGroup}{User-defined number of groups used for calculating calibration measures}

\item{nboot}{The number of bootstrapped samples used for calculating the predictive performance}
}
\description{
Develop risk prediction models with internal validation using bootstrap method
}
