% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxRiskModels.R
\name{coxModelEvaluation}
\alias{coxModelEvaluation}
\title{Model evaluation using established risk prediction models}
\usage{
coxModelEvaluation(
  dat = datTrainZ,
  time,
  status,
  Xs,
  S0,
  betaX,
  meanXB,
  predictYear,
  riskThreshold = NULL,
  calGroup,
  nboot = 100
)
}
\arguments{
\item{dat}{Data set used or the validation cohort}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{Xs}{Candidate predictors}

\item{S0}{The baseline survival probability}

\item{betaX}{The coefficients used in the risk prediction model}

\item{meanXB}{The means of candidate predictors}

\item{predictYear}{The t-year predicted risk}

\item{riskThreshold}{User-defined thresholds used for calculating calibration measures}

\item{calGroup}{User-defined number of groups used for calculating calibration measures}

\item{nboot}{The number of bootstrapped samples used for calculating the predictive performance}
}
\description{
Model evaluation using established risk prediction models
}
