% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRsRiskModels.R
\name{CRsModelEvaluation}
\alias{CRsModelEvaluation}
\title{Competing risk prediction models' evaluation}
\usage{
CRsModelEvaluation(cox.fit, dat, time, status, Yeoi, predictYear, nboot = 100)
}
\arguments{
\item{cox.fit}{The fitted model using either \code{rms::cph()} or \code{riskRegression::CSC()}}

\item{dat}{Data set or the derivation cohort}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{Yeoi}{Event value of interest}

\item{predictYear}{Predicted year of developing disease's risk}

\item{nboot}{The number of bootstrapped samples}
}
\description{
Competing risk prediction models' evaluation
}
