% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRsRiskModels.R
\name{CRsModelSelection}
\alias{CRsModelSelection}
\title{Competing risk prediction models with predictors selections}
\usage{
CRsModelSelection(
  Xs.list,
  datTrain,
  time,
  status,
  Yeoi,
  predictYear,
  nboot = 100,
  datValid
)
}
\arguments{
\item{Xs.list}{Candidate predictors list considered in developing risk models}

\item{datTrain}{Data set or the derivation cohort}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{Yeoi}{Event value of interest}

\item{predictYear}{Predicted year of developing disease's risk}

\item{nboot}{The number of bootstrapped samples}

\item{datValid}{Data set or the validation cohort}
}
\description{
Competing risk prediction models with predictors selections
}
