% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competingRiskModels.R, R/utilities.R
\name{datZScale}
\alias{datZScale}
\title{data scaling}
\usage{
datZScale(dat, time, status, Xs, newdat)

datZScale(dat, time, status, Xs, newdat)
}
\arguments{
\item{dat}{Data set or the derivation cohort}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{Xs}{Candidate predictors}

\item{newdat}{New data set or the external validation cohort}
}
\description{
data scaling

data scaling
}
