% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCal}
\alias{getCal}
\title{Calculate the calibration measures}
\usage{
getCal(dat, time, status, riskPred, predictYear)
}
\arguments{
\item{dat}{Data set}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{riskPred}{Predicted risk at t-year follow-up}

\item{predictYear}{Predicted year of developing disease's risk}
}
\description{
Calculate the calibration measures
}
\keyword{internal}
