% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCalSlopeInt}
\alias{getCalSlopeInt}
\title{Calcuate the calibration slope}
\usage{
getCalSlopeInt(cox.fit, dat, time, status, Yeoi, predictYear)
}
\arguments{
\item{cox.fit}{The fitted model using either \code{rms::cph()} or \code{riskRegression::CSC()}}

\item{dat}{Data set or the derivation cohort}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{Yeoi}{Event value of interest}

\item{predictYear}{Predicted year of developing disease's risk}
}
\description{
Calcuate the calibration slope
}
