% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competingRiskModels.R
\name{getCIF}
\alias{getCIF}
\title{Calculate the cumulative incidence function in the presence of competing risks}
\usage{
getCIF(dat, time, status, Xs, Yeoi, Ycrs, id, timePoints = c(1:5))
}
\arguments{
\item{dat}{Data set used in the analysis}

\item{time}{Time to event}

\item{status}{Status at the end of study}

\item{Xs}{Predictors used in the study}

\item{Yeoi}{Value of \code{status} indicate the event of interest}

\item{Ycrs}{Value of \code{status} indicate the competing risks}

\item{id}{ID for each individual}

\item{timePoints}{Time points of interest for calculating cumulative incidence}
}
\value{
A data frame with the estimated cumulative incidence rate at time points of interest
}
\description{
This function is used to calculate the cumulative incidence function based on the time-varying weights that corrects for right censored and left truncated data via a counting process format data. With the counting process format data, sub-distribution proportional hazards model can be performed. More details for the time-varying weights can be found at \code{\link[mstate]{crprep}}.
}
\examples{
library(riskModel)
head(rdata)
getCIF(dat=rdata,time="time",
       status="status_num",Yeoi=1,Ycrs=2,id="id",
       Xs=c("age","size"),timePoints=1:5)

}
\seealso{
\code{\link[mstate]{crprep}}, \code{\link[survival]{survfit}}
}
