% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getNRIIDI}
\alias{getNRIIDI}
\title{Calculate reclassification measures of NRI and IDI}
\usage{
getNRIIDI(dat, time, status, predrisk1, predrisk2, cutoff)
}
\arguments{
\item{dat}{Data set used}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{predrisk1}{Predicted risk from Model 1}

\item{predrisk2}{Predicted risk from Model 2}

\item{cutoff}{Cutoffs used for caculating NRI and IDI}
}
\description{
Calculate reclassification measures of NRI and IDI
}
