% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getSTDCA}
\alias{getSTDCA}
\title{Conduct decision curve analysis}
\usage{
getSTDCA(
  cox.fit,
  dat,
  time,
  status,
  Yeoi,
  predictYear,
  xstart = 0.01,
  xstop = 0.99,
  xby = 0.01,
  ymin = -0.05,
  probability = NULL,
  harm = NULL,
  interventionper = 100,
  smooth = FALSE,
  loess.span = 0.1,
  cmprsk = TRUE
)
}
\arguments{
\item{cox.fit}{The fitted model using either \code{rms::cph()} or \code{riskRegression::CSC()}}

\item{dat}{Data set or the derivation cohort}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{Yeoi}{Event value of interest}

\item{predictYear}{Predicted year of developing disease's risk}

\item{xstart}{The start value of X-axis}

\item{xstop}{The end value of X-axis}

\item{xby}{The step by the X-axis}

\item{ymin}{The smallest value of Y-axis}

\item{probability}{The Probability}

\item{harm}{The cost of false positive results}

\item{interventionper}{By default, \code{interventionper=100}}

\item{smooth}{Indicator of whether smoothed the net benefit curve. By default, \code{smooth=FALSE}}

\item{loess.span}{The bin of loess()}

\item{cmprsk}{Indicator of whether considering competing risk. By default, \code{cmprsk=FALSE}}
}
\description{
Conduct decision curve analysis
}
