% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GNDCalibrate}
\alias{GNDCalibrate}
\title{Conduct the GND Chi-squre test}
\usage{
GNDCalibrate(pred, tvar, out, cens.t, groups, adm.cens)
}
\arguments{
\item{pred}{Predicted risk}

\item{tvar}{Survival time}

\item{out}{Value for recording the event}

\item{cens.t}{Value for recording the censoring}

\item{groups}{Groups assigned to each observation}

\item{adm.cens}{End of follow-up time}
}
\description{
Conduct the GND Chi-squre test
}
\keyword{internal}
