% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotGroupCalibration}
\alias{plotGroupCalibration}
\title{Plot the grouped calibration plot}
\usage{
plotGroupCalibration(
  cox.fit,
  dat,
  time,
  status,
  Yeoi,
  predictYear,
  riskThreshold,
  maxRisk = 0.3
)
}
\arguments{
\item{cox.fit}{The fitted model using either \code{rms::cph()} or \code{riskRegression::CSC()}}

\item{dat}{Data set or the derivation cohort}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{Yeoi}{Event value of interest}

\item{predictYear}{Predicted year of developing disease's risk}

\item{riskThreshold}{User-defined thresholds used for calculating calibration measures}

\item{maxRisk}{The maximum value of the predicted risk}
}
\description{
Plot the grouped calibration plot
}
