% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotSTDCA}
\alias{plotSTDCA}
\title{Plot the decision curve}
\usage{
plotSTDCA(getSTDCA.fit, ylim = c(-0.1, 0.2))
}
\arguments{
\item{getSTDCA.fit}{The fitted results from \code{getSTDCA()}}

\item{ylim}{The min and max value of Y-axis}
}
\description{
Plot the decision curve
}
