% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotXsRisk}
\alias{plotXsRisk}
\title{Plot the predicted vs observed t-year risk}
\usage{
plotXsRisk(cox.fit, dat, Yeoi, predictYear, x, xlim = NULL, factorX = FALSE)
}
\arguments{
\item{cox.fit}{The fitted model using either \code{rms::cph()} or \code{riskRegression::CSC()}}

\item{dat}{Data set used}

\item{Yeoi}{Event value of interest}

\item{predictYear}{Predicted year of developing disease's risk}

\item{x}{The candidate predictor of interest}

\item{xlim}{The x-axis}

\item{factorX}{Indicator of categorical predictor. By default, \code{factorX=FALSE}.}
}
\description{
Plot the predicted vs observed t-year risk
}
