% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Rsq}
\alias{Rsq}
\title{Calculate R2}
\usage{
Rsq(lp, time, status, ties = TRUE)
}
\arguments{
\item{lp}{Linear predictors}

\item{time}{Survival time}

\item{status}{Event status at the end of follow-up}

\item{ties}{Allow for tied time. By default, \code{ties=TRUE}}
}
\description{
Calculate R2
}
\keyword{internal}
