% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datExample.R
\docType{data}
\name{vdata}
\alias{vdata}
\title{Example validation data set}
\format{
\code{vdata} contains
\describe{
  \item{id}{ID}
  \item{status}{Status at the end of study}
  \item{time}{Time to event}
  \item{age}{Age at baseline}
  \item{size}{Size of the tumor}
  \item{ncat}{Stage}
  \item{hr_status}{Status of HER2 expression}
  \item{status_num}{Indicates of status}
}
}
\source{
<https://github.com/survival-lumc/ValidationCompRisks/tree/main/Data>
}
\usage{
vdata
}
\description{
Example validation data set
}
\keyword{datasets}
